fix(receive): #1407 continue to  show btc invoice if instant payment is not confirmed
